[#183] refactor french word vector
