+++ src/ml_service/feature_extraction/pre_processing/precedent_model.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 class PrecedentModel:
     def __init__(self):
         self.dict = {
