+++ src/ml_service/feature_extraction/pre_processing/regex_replacer.py
@@ -4,7 +4,7 @@ import string
 from nltk.tokenize import word_tokenize
 from nltk.tokenize.moses import MosesDetokenizer
 
-from feature_extraction.pre_processing.word_vectors.vectors import FrenchVectors
+from feature_extraction.pre_processing.word_vectors.french_vector import FrenchVectors
 
 
 class RegexReplacer():
